Make as many global variables local as possible
Effigy needed to save a value in the Joker itself.
I've shown the usual method for doing this.

Some jokers required global variables and their code
could not be easily changed. Added comments to those

Using global variables in a Joker is almost always wrong
and will lead to these bugs:
- Multiples of that Joker clobber each other's variables
- State persists after starting a new run, and wrongly does not persist after loading a saved game
- Will conflict with other mods